feat(module): [user] move to plugins
