--- examples/next-starter/src/modules/user/entities/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from './IUser.d';
