--- examples/next-starter/src/modules/user/services/interfaces/UserRequestQueryPayload.d.ts
@@ -1,3 +1,0 @@
-import { HttpModel } from '@onr/core';
-
-export interface UserRequestQueryPayload extends HttpModel.IRequestQueryPayload {}
