--- examples/next-starter/src/modules/user/services/interfaces/index.ts
@@ -1,5 +1,0 @@
-export * from './UserModel.d';
-export * from './UserRequestPayload.d';
-export * from './UserRequestQueryPayload.d';
-export * from './UserResponse.d';
-export * from './UsersResponse.d';
