+++ plugins/user/src/entities/interfaces/IUser.ts
