+++ plugins/user/src/services/interfaces/UserModel.ts
@@ -1,0 +1,23 @@
+import { AnyNsRecord } from 'dns';
+import { UserRequestPayload, UserResponse, UsersResponse } from '.';
+
+export interface GetUsersPayload {
+  params?: AnyNsRecord;
+}
+
+export interface GetUserPayload extends GetUsersPayload {
+  userId: number;
+}
+
+export interface CreateUserPayload extends UserRequestPayload {}
+
+export interface UpdateUserPayload extends UserRequestPayload {
+  userId: number;
+}
+
+export interface DeleteUserPayload {
+  userId: number;
+}
+
+export interface GetUserResponse extends UserResponse {}
+export interface GetUsersResponse extends UsersResponse {}
