+++ plugins/user/src/services/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './UserModel';
+export * from './UserRequestPayload';
+export * from './UserResponse';
+export * from './UsersResponse';
